[REF] l10n_br_fiscal: merge document.line.mixin.methods into document.line.mixin
